backend/requirements.txt
groq
fastapi
uvicorn
python-dotenv
python-jose 
passlib[bcrypt]
bcrypt<4.1
ollama
PyPDF2
sentence_transformers
numpy
scikit-learn

pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0

PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6

chromadb>=0.5.3